Refactor genetic algorithm runners
