+++ jmetal/runner/singleobjective/generational_genetic_algorithm_running_as_a_thread.py
@@ -1,0 +1,37 @@
+import logging
+
+from jmetal.algorithm.singleobjective.evolutionaryalgorithm import GenerationalGeneticAlgorithm
+from jmetal.core.solution import BinarySolution, FloatSolution
+from jmetal.operator.crossover import SinglePoint, SBX
+from jmetal.operator.mutation import BitFlip, Polynomial
+from jmetal.operator.selection import BinaryTournament
+from jmetal.problem.singleobjectiveproblem import OneMax, Sphere
+
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+
+def gGARunningAsAThread() -> None:
+    variables = 10
+    problem = Sphere(variables)
+    algorithm = GenerationalGeneticAlgorithm[FloatSolution, FloatSolution](
+        problem,
+        population_size = 100,
+        max_evaluations = 25000,
+        mutation = Polynomial(1.0/variables, distribution_index=20),
+        crossover = SBX(1.0, distribution_index=20),
+        selection = BinaryTournament())
+
+    algorithm.start()
+    logger.info("Algorithm (running as a thread): " + algorithm.get_name())
+    logger.info("Problem: " + problem.get_name())
+
+    algorithm.join()
+    result = algorithm.get_result()
+    logger.info("Solution: " + str(result.variables))
+    logger.info("Fitness:  " + str(result.objectives[0]))
+    logger.info("Computing time: " + str(algorithm.total_computing_time))
+
+
+if __name__ == '__main__':
+    gGARunningAsAThread()
